# backend/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a.",
        "title": "Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Manager API Support",
            "url": "https://github.com/Inteli-College/2024-1B-T02-EC10-G04",
            "email": "gomedicine@inteli.edu.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/medicines": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all Medicines entities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medicines"
                ],
                "summary": "Retrieve all Medicines entities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.FindMedicineOutputDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new Medicine entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medicines"
                ],
                "summary": "Create a new Medicine entity",
                "parameters": [
                    {
                        "description": "Medicine entity to create",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMedicineInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMedicineOutputDTO"
                        }
                    }
                }
            }
        },
        "/medicines/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a Medicine entity by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medicines"
                ],
                "summary": "Retrieve a Medicine entity by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FindMedicineOutputDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a Medicine entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medicines"
                ],
                "summary": "Update a Medicine entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medicine entity to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMedicineInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FindMedicineOutputDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a Medicine entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medicines"
                ],
                "summary": "Delete a Medicine entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all Order entities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Retrieve all Order entities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.FindOrderOutputDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new Order entity and produce an event to Kafka",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new Order entity",
                "parameters": [
                    {
                        "description": "Order entity to create",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrdersInputDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders/collector": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all Order entities by Collector ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Retrieve all Order entities by Collector ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.FindOrderOutputDTO"
                            }
                        }
                    }
                }
            }
        },
        "/orders/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all Order entities by User ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Retrieve all Order entities by User ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.FindOrderOutputDTO"
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve an Order entity by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Retrieve an Order entity by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FindOrderOutputDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an Order entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Update an Order entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Order entity to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOrderInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOrderOutputDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an Order entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an Order entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateMedicineInputDTO": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stripe": {
                    "$ref": "#/definitions/entity.StripeType"
                }
            }
        },
        "dto.CreateMedicineOutputDTO": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stripe": {
                    "$ref": "#/definitions/entity.StripeType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.CreateOrdersInputDTO": {
            "type": "object",
            "required": [
                "medicine_ids",
                "priority",
                "pyxis_id"
            ],
            "properties": {
                "medicine_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "observation": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "pyxis_id": {
                    "type": "string"
                },
                "responsible_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.FindMedicineOutputDTO": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stripe": {
                    "$ref": "#/definitions/entity.StripeType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.FindOrderOutputDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "medicine": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Medicine"
                    }
                },
                "medicine_id": {
                    "type": "string"
                },
                "observation": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "pyxis_id": {
                    "type": "string"
                },
                "responsible": {
                    "$ref": "#/definitions/entity.User"
                },
                "responsible_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/entity.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateMedicineInputDTO": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stripe": {
                    "$ref": "#/definitions/entity.StripeType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateOrderInputDTO": {
            "type": "object",
            "properties": {
                "observation": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "responsible_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateOrderOutputDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "medicines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Medicine"
                    }
                },
                "observation": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "responsible": {
                    "$ref": "#/definitions/entity.User"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "$ref": "#/definitions/entity.User"
                }
            }
        },
...
        "entity.Medicine": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stripe": {
                    "$ref": "#/definitions/entity.StripeType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entity.StripeType": {
            "type": "string",
            "enum": [
                "red",
                "yellow",
                "black"
            ],
            "x-enum-varnames": [
                "StripeRed",
                "StripeYellow",
                "StripeBlack"
            ]
        },
        "entity.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "on_duty": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Type: Bearer token\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}